Added cnl_epochs
Messed around alot with matlab, didn't get that much physical work done,
but am getting more familiar with matlabs OOP.  created the epochs
class, didnt finish it all though yet. did a basic outline and tested
it. wrote getIndexInEpochs and tested it using the epochs class and it
worked.  changed the constructor for TimeSeries to be (timeline, signal)
instead of (signal,timeline) added the introdata set to the github for
easy access purposes. created CNLWickenden username for when we are at
wickenden, password is same as the computer password except there is a 1
after it.